Add multiple contracts to one endpoint
Added a second contract specification to an endpoint to demonstrate how
to have different responses come from a single endpoint based on the
request sent in.

The idea being to test positive and negative test scenerios. In this
case for most requests, we return a user, but for a specific id we
return 'No user found' instead, thus creating the ability to contract
test a failure / negative case.